backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pyodbc==5.1.0
alembic==1.13.0
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
cryptography==43.0.0
azure-storage-blob==12.23.0
python-multipart==0.0.12
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
pyyaml==6.0.2
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
